+++ data/src/main/java/com/gyleedev/data/database/dao/MessageDao.kt
@@ -1,4 +1,4 @@
-package com.gyleedev.chatchat.data.database.dao
+package com.gyleedev.data.database.dao
 
 import androidx.paging.PagingSource
 import androidx.room.Dao
@@ -6,7 +6,7 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Update
-import com.gyleedev.chatchat.data.database.entity.MessageEntity
+import com.gyleedev.data.database.entity.MessageEntity
 import kotlinx.coroutines.flow.Flow
 
 @Dao
