+++ data/src/main/java/com/gyleedev/data/database/dao/UserAndFavoriteDao.kt
@@ -1,4 +1,4 @@
-package com.gyleedev.chatchat.data.database.dao
+package com.gyleedev.data.database.dao
 
 import androidx.paging.PagingSource
 import androidx.room.Dao
@@ -7,8 +7,8 @@ import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Transaction
 import androidx.room.Update
-import com.gyleedev.chatchat.data.database.entity.UserAndFavoriteEntity
-import com.gyleedev.chatchat.data.database.entity.UserEntity
+import com.gyleedev.data.database.entity.UserAndFavoriteEntity
+import com.gyleedev.data.database.entity.UserEntity
 import com.gyleedev.domain.model.UserRelationState
 import kotlinx.coroutines.flow.Flow
