+++ data/src/main/java/com/gyleedev/data/repository/MessageRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package com.gyleedev.chatchat.data.repository
+package com.gyleedev.data.repository
 
 import android.net.Uri
 import android.os.Build
@@ -13,11 +13,11 @@ import com.google.firebase.database.DataSnapshot
 import com.google.firebase.database.DatabaseError
 import com.google.firebase.database.FirebaseDatabase
 import com.google.firebase.storage.FirebaseStorage
-import com.gyleedev.chatchat.data.database.dao.MessageDao
-import com.gyleedev.chatchat.data.database.entity.toEntity
-import com.gyleedev.chatchat.data.database.entity.toModel
-import com.gyleedev.chatchat.data.database.entity.toUpdateEntity
-import com.gyleedev.chatchat.data.preference.MyDataPreference
+import com.gyleedev.data.database.dao.MessageDao
+import com.gyleedev.data.database.entity.toEntity
+import com.gyleedev.data.database.entity.toModel
+import com.gyleedev.data.database.entity.toUpdateEntity
+import com.gyleedev.data.preference.MyDataPreference
 import com.gyleedev.domain.model.ChatRoomLocalData
 import com.gyleedev.domain.model.MessageData
 import com.gyleedev.domain.model.MessageSendState
